pyproject.toml
[project]
name = "trading-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Gajesh Naik",email = "26431906+Gajesh2007@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "hyperliquid-python-sdk (>=0.20.0,<0.21.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "web3 (>=7.14.0,<8.0.0)",
    "eth-abi (>=5.2.0,<6.0.0)",
    "eth-account (>=0.13.0,<1.0.0)",
    "aiohttp (>=3.13.1,<4.0.0)",
    "openai (>=2.5.0,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "rich (>=14.2.0,<15.0.0)",
    "pandas (>=2.0.0,<3.0.0)",
    "pandas-ta (>=0.3.14b0,<1.0.0)",
    "python-binance (>=1.0.19,<2.0.0)",
    "fastapi (>=0.115.0,<1.0.0)",
    "uvicorn (>=0.32.0,<1.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
